streamline server and client loops
I finally got around to reporting errors from the script
execution and with the previous work on bindings for ENet
types I can take out most of the logic from the loops.

This means that now network.js is in charge of the ENet logic
and with a bit of luck I can make another bindings library
for serializing and de-serializing data I receive from the ENet
loop.