correctly implement enet event pulling
There are two loops here. The first is just to make the thread spin
for as long as we need and the second is to pull out of the ENet
queue all the network events we can.

In the inner loop you need a fresh event structure for each ENet
event. The structure is filled using enet_host_service but here I
was reusing the same pointer after dispatching it to the event loop.

This will instead allocate a new struct after each dispatch.

Moreover when enet_host_service returns 0 breaking the inner loop
the event structure is unused and can be safely deallocated.